move LamportsError to solana-instruction
